*: move basic functions and definitions to model package, to make constructing dag request more easier.
